pyproject.toml
[tool.poetry]
name = "NXLog task"
version = "0.1.0"
description = "Automation Framework using Python, Playwright, and Pytest"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.0"
playwright = "^1.50"
pytest-playwright = "^0.5"
allure-pytest = "^2.13.2"
#python-dotenv = "^1.0"

#[tool.poetry.group.dev.dependencies]
#black = "^24.0"
#flake8 = "^7.1"
#pytest-html = "^4.1"
#pytest-xdist = "^3.6"